split the test_structure into smaller test files
