+++ tests/dflowfm/structure/test_structure.py
@@ -1,25 +1,15 @@
 import inspect
-from pathlib import Path
 from typing import Any, Dict, List, Union
 
 import pytest
 from pydantic import ValidationError
 from unittest.mock import MagicMock
-from hydrolib.core.dflowfm.bc.models import ForcingModel
 from hydrolib.core.dflowfm.friction.models import FrictionType
 from hydrolib.core.dflowfm.ini.parser import Parser, ParserConfig
 from hydrolib.core.dflowfm.structure.models import (
-    Bridge,
     Compound,
     Culvert,
-    CulvertSubType,
-    Dambreak,
-    DambreakAlgorithm,
     FlowDirection,
-    GateOpeningHorizontalDirection,
-    GeneralStructure,
-    LongCulvert,
-    Orientation,
     Orifice,
     Pump,
     Structure,
@@ -27,14 +17,11 @@ from hydrolib.core.dflowfm.structure.models import (
     UniversalWeir,
     Weir,
 )
-from hydrolib.core.dflowfm.tim.models import TimModel
 
 from tests.utils import (
     WrapperTest,
-    create_temp_file,
     invalid_test_data_dir,
     test_data_dir,
-    test_input_dir,
 )
 
 uniqueid_str = "Unique structure id (max. 256 characters)."
@@ -252,12 +239,6 @@ def _get_allowedflowdir_cases() -> List:
     ]
 
 
-def _get_orientation_cases() -> List:
-    return [
-        ("Positive", "positive"),
-        ("NEGATIVE", "negative"),
-    ]
-
 
 @pytest.mark.parametrize(
     "input,expected",
@@ -343,198 +324,16 @@ def test_culvert_parses_subtype_case_insensitive(input, expected):
     assert structure.subtype == expected
 
 
-def create_structure_values(type: str) -> dict:
+def create_structure_values(struture_type: str) -> dict:
     """Create a dict with example field values for all common structure attributes."""
     return dict(
         id="structure_id",
         name="structure_name",
-        type=type,
+        type=struture_type,
         branchid="branch_id",
         chainage="1.23",
     )
 
-
-class TestBridge:
-    """
-    Wrapper class to test all the methods and subclasses in:
-    hydrolib.core.dflowfm.structure.models.py Bridge class
-    """
-
-    def test_create_a_bridge_from_scratch(self):
-        bridge = Bridge(
-            id="b003",
-            name="B003",
-            branchid="B1",
-            chainage=5.0,
-            allowedflowdir=FlowDirection.both,
-            csdefid="W_980.1S_0",
-            shift=-1.23,
-            inletlosscoeff=1,
-            outletlosscoeff=1,
-            frictiontype=FrictionType.strickler,
-            friction=70,
-            length=100,
-            comments=Bridge.Comments(
-                name="B stands for Bridge, 003 because we expect to have at most 999 weirs"
-            ),
-        )
-
-        assert isinstance(
-            bridge, Structure
-        ), "Bridge should also be an instance of a Structure"
-
-        assert bridge.id == "b003"
-        assert bridge.name == "B003"
-        assert bridge.branchid == "B1"
-        assert bridge.chainage == pytest.approx(5.0)
-        assert bridge.allowedflowdir == FlowDirection.both
-        assert bridge.csdefid == "W_980.1S_0"
-        assert bridge.shift == -1.23
-        assert bridge.inletlosscoeff == 1
-        assert bridge.outletlosscoeff == 1
-        assert bridge.frictiontype == FrictionType.strickler
-        assert bridge.friction == 70
-        assert bridge.length == 100
-        assert (
-            bridge.comments.name
-            == "B stands for Bridge, 003 because we expect to have at most 999 weirs"
-        )
-
-        assert bridge.comments.id == uniqueid_str
-
-    def test_bridge_construction_with_parser(self):
-        parser = Parser(ParserConfig())
-
-        input_str = inspect.cleandoc(
-            """
-            [Structure]
-            type = bridge                   # Structure type; must read bridge
-            id = RS1-KBR31                  # Unique structure id (max. 256 characters).
-            name = RS1-KBR31name            # Given name in the user interface.
-            branchid = riv_RS1_264          # (optional) Branch on which the structure is located.
-            chainage = 104.184              # (optional) Chainage on the branch (m)
-            allowedFlowDir = both           # Possible values: both, positive, negative, none.
-            csDefId = W_980.1S_0            # Id of Cross-Section Definition.
-            shift = 0.0                     # Vertical shift of the cross section definition [m]. Defined positive upwards.
-            inletLossCoeff = 1              # Inlet loss coefficient [-], ?_i.
-            outletLossCoeff = 1
-            frictionType = Strickler        # Friction type
-            friction = 70
-            length = 9.75
-            """
-        )
-
-        for line in input_str.splitlines():
-            parser.feed_line(line)
-
-        document = parser.finalize()
-
-        wrapper = WrapperTest[Bridge].parse_obj({"val": document.sections[0]})
-        bridge = wrapper.val
-
-        assert isinstance(
-            bridge, Structure
-        ), "Bridge should also be an instance of a Structure"
-
-        assert bridge.id == "RS1-KBR31"
-        assert bridge.name == "RS1-KBR31name"
-        assert bridge.branchid == "riv_RS1_264"
-        assert bridge.chainage == pytest.approx(104.184)
-        assert bridge.type == "bridge"
-        assert bridge.allowedflowdir == FlowDirection.both
-        assert bridge.csdefid == "W_980.1S_0"
-        assert bridge.shift == pytest.approx(0.0)
-        assert bridge.inletlosscoeff == 1
-        assert bridge.outletlosscoeff == 1
-        assert bridge.frictiontype == FrictionType.strickler
-        assert bridge.friction == 70
-        assert bridge.length == pytest.approx(9.75)
-
-    def test_bridge_with_unknown_parameter_is_ignored(self):
-        parser = Parser(ParserConfig())
-
-        input_str = inspect.cleandoc(
-            """
-            [Structure]
-            type = bridge                   # Structure type; must read bridge
-            id = RS1-KBR31                  # Unique structure id (max. 256 characters).
-            name = RS1-KBR31name            # Given name in the user interface.
-            branchid = riv_RS1_264          # (optional) Branch on which the structure is located.
-            chainage = 104.184              # (optional) Chainage on the branch (m)
-            allowedFlowDir = both           # Possible values: both, positive, negative, none.
-            csDefId = W_980.1S_0            # Id of Cross-Section Definition.
-
-            # ----------------------------------------------------------------------
-            unknown           = 10.0        # A deliberately added unknown property
-            # ----------------------------------------------------------------------
-
-            shift = 0.0                     # Vertical shift of the cross section definition [m]. Defined positive upwards.
-            inletLossCoeff = 1              # Inlet loss coefficient [-], ?_i.
-            outletLossCoeff = 1
-            frictionType = Strickler        # Friction type
-            friction = 70
-            length = 9.75
-            """
-        )
-
-        for line in input_str.splitlines():
-            parser.feed_line(line)
-
-        document = parser.finalize()
-
-        wrapper = WrapperTest[Bridge].parse_obj({"val": document.sections[0]})
-        bridge = wrapper.val
-
-        assert bridge.dict().get("unknown") is None  # type: ignore
-
-        assert bridge.id == "RS1-KBR31"
-        assert bridge.name == "RS1-KBR31name"
-        assert bridge.branchid == "riv_RS1_264"
-        assert bridge.chainage == pytest.approx(104.184)
-        assert bridge.type == "bridge"
-        assert bridge.allowedflowdir == FlowDirection.both
-        assert bridge.csdefid == "W_980.1S_0"
-        assert bridge.shift == pytest.approx(0.0)
-        assert bridge.inletlosscoeff == 1
-        assert bridge.outletlosscoeff == 1
-        assert bridge.frictiontype == FrictionType.strickler
-        assert bridge.friction == 70
-        assert bridge.length == pytest.approx(9.75)
-
-    def test_bridge_with_missing_required_parameters(self):
-        parser = Parser(ParserConfig())
-
-        input_str = inspect.cleandoc(
-            """
-            [Structure]
-            type = bridge                   # Structure type; must read bridge
-            id = RS1-KBR31                  # Unique structure id (max. 256 characters).
-            name = RS1-KBR31name            # Given name in the user interface.
-            branchid = riv_RS1_264          # (optional) Branch on which the structure is located.
-            chainage = 104.184              # (optional) Chainage on the branch (m)
-            allowedFlowDir = both           # Possible values: both, positive, negative, none.
-            # csDefId = W_980.1S_0            # Id of Cross-Section Definition.
-            # shift = 0.0                     # Vertical shift of the cross section definition [m]. Defined positive upwards.
-            # inletLossCoeff = 1              # Inlet loss coefficient [-], ?_i.
-            # outletLossCoeff = 1
-            # frictionType = Strickler        # Friction type
-            # friction = 70
-            # length = 9.75
-            """
-        )
-
-        for line in input_str.splitlines():
-            parser.feed_line(line)
-
-        document = parser.finalize()
-
-        expected_message = "7 validation errors for WrapperTest[Bridge]"
-        with pytest.raises(ValueError) as exc_err:
-            wrapper = WrapperTest[Bridge].model_validate({"val": document.sections[0]})
-            bridge = wrapper.val
-        assert expected_message in str(exc_err.value)
-
-        assert "bridge" not in locals()  # Bridge structure should not have been created
 
 
 class TestStructure:
@@ -869,1592 +668,3 @@ class TestStructure:
                 == f"Expected at least 2 coordinates, but only {n_coords} declared."
             )
 
-
-class TestDambreakAlgorithm:
-    """
-    Wrapper class to test all the methods in:
-    hydrolib.core.dflowfm.structure.models.py DambreakAlgorithm enum class.
-    """
-
-    @pytest.mark.parametrize(
-        "enum_value, enum_description",
-        [
-            pytest.param(1, "van der Knaap, 2000"),
-            pytest.param(2, "Verheij-van der Knaap, 2002"),
-            pytest.param(3, "Predefined time series, dambreakLevelsAndWidths"),
-        ],
-    )
-    def test_get_enum_as_str_returns_description(
-        self, enum_value: int, enum_description: str
-    ):
-        assert DambreakAlgorithm(enum_value).description == enum_description
-
-
-class DambreakTestCases:
-    """Just a wrapper so it can be referenced from other classes."""
-
-    check_location_err = "Specify location either by setting `num/x/yCoordinates` or `polylinefile` fields for a Dambreak structure."
-    check_upstream_waterlevel_location_err = "Either `waterLevelUpstreamNodeId` should be specified or `waterLevelUpstreamLocationX` and `waterLevelUpstreamLocationY`."
-    check_downstream_waterlevel_location_err = "Either `waterLevelDownstreamNodeId` should be specified or `waterLevelDownstreamLocationX` and `waterLevelDownstreamLocationY`."
-    too_few_coords = "Expected at least 2 coordinates, but only {} declared."
-    mismatch_coords = (
-        "Expected {} coordinates, given {} for xCoordinates and {} for yCoordinates."
-    )
-
-
-class TestDambreak:
-    """
-    Wrapper class to test all the methods and sublcasses in:
-    hydrolib.core.dflowfm.structure.models.py Dambreak class.
-    """
-
-    @pytest.fixture
-    def default_dambreak_values(self) -> dict:
-        return dict(
-            id="NLNJ-2021",
-            name="NederlandNaranjito2021",
-            type="dambreak",
-            startlocationx=4.2,
-            startlocationy=2.4,
-            algorithm=1,
-            crestlevelini=1,
-            breachwidthini=24,
-            crestlevelmin=42,
-            t0=2.2,
-            timetobreachtomaximumdepth=4.4,
-            f1=22.4,
-            f2=44.2,
-            ucrit=44.22,
-            waterlevelupstreamnodeid="anUpstreamNodeId",
-            waterleveldownstreamnodeid="aDownstreamNodeId",
-        )
-
-    @pytest.fixture
-    def valid_dambreak_values(self, default_dambreak_values: dict) -> dict:
-        coordinates_dict = dict(
-            numcoordinates=2,
-            xcoordinates=[4.2, 2.4],
-            ycoordinates=[2.4, 4.2],
-        )
-        return {**default_dambreak_values, **coordinates_dict}
-
-    @pytest.mark.parametrize(
-        "location_dict, err_mssg",
-        [
-            pytest.param(
-                dict(),
-                DambreakTestCases.check_location_err,
-                id="No values.",
-            ),
-            pytest.param(
-                dict(numcoordinates=None),
-                DambreakTestCases.check_location_err,
-                id="Missing x_y_coordinates",
-            ),
-            pytest.param(
-                dict(numcoordinates=None, ycoordinates=None),
-                DambreakTestCases.check_location_err,
-                id="Missing xcoordinates",
-            ),
-            pytest.param(
-                dict(numcoordinates=None, xcoordinates=None),
-                DambreakTestCases.check_location_err,
-                id="Missing ycoordinates",
-            ),
-            pytest.param(
-                dict(numcoordinates=0, xcoordinates=None, ycoordinates=None),
-                DambreakTestCases.check_location_err,
-                id="None coordinates.",
-            ),
-            pytest.param(
-                dict(numcoordinates=0, xcoordinates=[], ycoordinates=[]),
-                DambreakTestCases.too_few_coords.format(0),
-                id="Empty list coordinates.",
-            ),
-            pytest.param(
-                dict(numcoordinates=1, xcoordinates=[], ycoordinates=[]),
-                DambreakTestCases.too_few_coords.format(1),
-                id="One coordinate.",
-            ),
-            pytest.param(
-                dict(numcoordinates=2, xcoordinates=[4.2], ycoordinates=[]),
-                DambreakTestCases.mismatch_coords.format(2, 1, 0),
-                id="Mismatch coordinates.",
-            ),
-        ],
-    )
-    def test_given_invalid_location_raises_validation_error(
-        self,
-        location_dict: dict,
-        err_mssg: str,
-        default_dambreak_values: dict,
-    ):
-        test_values = {**default_dambreak_values, **location_dict}
-
-        with pytest.raises(ValueError) as exc_err:
-            Dambreak(**test_values)
-
-        assert err_mssg in str(exc_err.value)
-
-    def test_given_valid_values_creates_dambreak(self, valid_dambreak_values: dict):
-        dambreak = Dambreak(**valid_dambreak_values)
-        self.validate_valid_default_dambreak(dambreak)
-
-    def test_given_structure_text_with_num_x_y_coordinates_parses_structure(self):
-
-        structure_text = inspect.cleandoc(
-            """
-            [Structure]
-            type = dambreak                 # Structure type; must read dambreak
-            id = NLNJ-2021                  # Unique structure id (max. 256 characters).
-            name = NederlandNaranjito2021   # Given name in the user interface.
-            numCoordinates = 2              # Number of values in xCoordinates and yCoordinates. This value should be greater or equal 2.
-            xCoordinates = 4.2 2.4       # x-coordinates of the link selection polygon.
-            yCoordinates = 2.4 4.2       # y-coordinates of the link selection polygon.
-            startLocationX = 4.2            # x-coordinate of breach starting point.
-            startLocationY = 2.4            # y-coordinate of breach starting point.
-            algorithm = 1                   # Breach growth algorithm.
-
-            crestLevelIni = 1               # Initial breach level.
-            breachWidthIni = 24             # Initial breach width.
-            crestLevelMin = 42              # Minimal breach level.
-            t0 = 2.2                        # Breach start time.
-            timeToBreachToMaximumDepth = 4.4    # tphase 1.
-            f1 = 22.4                       # Factor f1.
-            f2 = 44.2                       # Factor f2.
-            uCrit = 44.22                   # Critical flow velocity uc for erosion [m/s].
-            waterLevelUpstreamNodeId = anUpstreamNodeId # Node Id of custom upstream water level point.
-            waterLevelDownstreamNodeId = aDownstreamNodeId # Node Id of custom downstream water level point.
-            """
-        )
-
-        dambreak_obj = self.parse_dambreak_from_text(structure_text)
-        self.validate_valid_default_dambreak(dambreak_obj)
-
-    def test_given_structure_text_with_polylinefile_parses_structure(self):
-        structure_text = inspect.cleandoc(
-            """
-            [structure]
-            type                       = dambreak  
-            id                         = dambreak  
-            polylinefile               = dambreak2ddrybreach.pli
-            startLocationX             = 1.2 
-            startLocationY             = 4.0
-            algorithm                  = 3             # 1 VdKnaap ,2 Verheij-vdKnaap
-            crestLevelIni              = 0.4
-            breachWidthIni             = 1
-            crestLevelMin              = 0.2
-            timeToBreachToMaximumDepth = 0.1           #in seconds 
-            f1                         = 1
-            f2                         = 1
-            ucrit                      = 0.001
-            t0                         = 0.0001        # make it a boolean
-            dambreakLevelsAndWidths    = {}  #used only in algorithm 3            
-            waterLevelUpstreamLocationX = 1.2 # x-coordinate of custom upstream water level point.
-            waterLevelUpstreamLocationY = 2.3 # y-coordinate of custom upstream water level point.
-            waterLevelDownstreamLocationX = 3.4 # x-coordinate of custom downstream water level point.
-            waterLevelDownstreamLocationY = 4.5 # y-coordinate of custom downstream water level point.
-            """
-        )
-
-        with create_temp_file("", "dambreak.tim") as tim_file:
-            structure_text = structure_text.format(tim_file)
-
-            dambreak_obj = self.parse_dambreak_from_text(structure_text)
-
-        assert dambreak_obj
-        assert isinstance(dambreak_obj, Structure)
-        assert dambreak_obj.type == "dambreak"
-        assert dambreak_obj.id == "dambreak"
-        assert dambreak_obj.polylinefile.filepath == Path("dambreak2ddrybreach.pli")
-        assert dambreak_obj.startlocationx == pytest.approx(1.2)
-        assert dambreak_obj.startlocationy == pytest.approx(4.0)
-        assert dambreak_obj.algorithm == 3
-        assert dambreak_obj.crestlevelini == pytest.approx(0.4)
-        assert dambreak_obj.breachwidthini == 1
-        assert dambreak_obj.crestlevelmin == pytest.approx(0.2)
-        assert dambreak_obj.timetobreachtomaximumdepth == pytest.approx(0.1)
-        assert dambreak_obj.f1 == 1
-        assert dambreak_obj.f2 == 1
-        assert dambreak_obj.ucrit == pytest.approx(0.001)
-        assert dambreak_obj.t0 == pytest.approx(0.0001)
-        assert dambreak_obj.dambreaklevelsandwidths.filepath.name == "dambreak.tim"
-        assert dambreak_obj.waterlevelupstreamlocationx == pytest.approx(1.2)
-        assert dambreak_obj.waterlevelupstreamlocationy == pytest.approx(2.3)
-        assert dambreak_obj.waterleveldownstreamlocationx == pytest.approx(3.4)
-        assert dambreak_obj.waterleveldownstreamlocationy == pytest.approx(4.5)
-        assert dambreak_obj.waterlevelupstreamnodeid == None
-        assert dambreak_obj.waterleveldownstreamnodeid == None
-
-    def parse_dambreak_from_text(self, structure_text: str) -> Dambreak:
-        """
-        Method just to simplify how the tests can parse a dambreak without having to
-        repeat the same code for each test.
-
-        Args:
-            structure_text (str): Text containing Dambreak structure.
-
-        Returns:
-            Dambreak: Parsed object.
-        """
-        parser = Parser(ParserConfig())
-        for line in structure_text.splitlines():
-            parser.feed_line(line)
-        document = parser.finalize()
-
-        return WrapperTest[Dambreak].model_validate({"val": document.sections[0]}).val
-
-    def validate_valid_default_dambreak(self, dambreak: Dambreak):
-        """
-        Method to validate the default (valid) test case for a Dambreak.
-
-        Args:
-            dambreak (Dambreak): Instance that needs to be verified for its expected values.
-        """
-
-        assert isinstance(dambreak, Structure), "A dambreak should be a structure."
-        assert dambreak.id == "NLNJ-2021"
-        assert dambreak.name == "NederlandNaranjito2021"
-        assert dambreak.type == "dambreak"
-        assert dambreak.startlocationx == pytest.approx(4.2)
-        assert dambreak.startlocationy == pytest.approx(2.4)
-        assert dambreak.algorithm == 1
-        assert dambreak.crestlevelini == 1
-        assert dambreak.breachwidthini == 24
-        assert dambreak.crestlevelmin == 42
-        assert dambreak.t0 == pytest.approx(2.2)
-        assert dambreak.timetobreachtomaximumdepth == pytest.approx(4.4)
-        assert dambreak.f1 == pytest.approx(22.4)
-        assert dambreak.f2 == pytest.approx(44.2)
-        assert dambreak.ucrit == pytest.approx(44.22)
-        assert dambreak.numcoordinates == 2
-        assert dambreak.xcoordinates == [4.2, 2.4]
-        assert dambreak.ycoordinates == [2.4, 4.2]
-        assert dambreak.waterlevelupstreamlocationx == None
-        assert dambreak.waterlevelupstreamlocationy == None
-        assert dambreak.waterleveldownstreamlocationx == None
-        assert dambreak.waterleveldownstreamlocationy == None
-        assert dambreak.waterlevelupstreamnodeid == "anUpstreamNodeId"
-        assert dambreak.waterleveldownstreamnodeid == "aDownstreamNodeId"
-
-    class TestValidateAlgorithm:
-        """
-        Wrapper to validate all paradigms of validate_algorithm
-        """
-
-        @pytest.mark.parametrize(
-            "value",
-            [pytest.param("", id="Empty string."), pytest.param(None, id="None")],
-        )
-        def test_given_not_int_raises_value_error(self, value: Any):
-            with pytest.raises(ValueError) as exc_err:
-                Dambreak.validate_algorithm(value)
-            assert (
-                str(exc_err.value) == "Dambreak algorithm value should be of type int."
-            )
-
-        @pytest.mark.parametrize(
-            "value",
-            [pytest.param(0), pytest.param(4), pytest.param(-1)],
-        )
-        def test_given_value_out_of_range_raises_value_error(self, value: int):
-            with pytest.raises(ValueError) as exc_err:
-                Dambreak.validate_algorithm(value)
-            assert str(exc_err.value) == "Dambreak algorithm value should be 1, 2 or 3."
-
-        @pytest.mark.parametrize(
-            "value",
-            [pytest.param(1), pytest.param(2), pytest.param(3)],
-        )
-        def test_given_valid_value_returns_value(self, value: int):
-            assert Dambreak.validate_algorithm(value) == value
-
-    class TestValidateDambreakLevelsAndWidths:
-        """
-        Wrapper to validate all paradigms of
-        validate_dambreak_leels_and_widths
-        """
-
-        algorithm_values = [(-1), (0), (1), (2), (4)]
-
-        @pytest.mark.parametrize(
-            "algorithm_value",
-            algorithm_values,
-        )
-        @pytest.mark.parametrize(
-            "field_value",
-            [
-                pytest.param("", id="Empty string"),
-                pytest.param(Path(""), id="Empty Path"),
-                pytest.param("JustAValue", id="String value"),
-                pytest.param(Path("JustAValue"), id="Path value"),
-            ],
-        )
-        def test_given_field_value_but_no_algorithm_3_raises_value_error(
-            self, field_value: str, algorithm_value: int
-        ):
-            with pytest.raises(ValueError) as exc_err:
-                Dambreak.validate_dambreak_levels_and_widths(
-                    field_value,
-                    dict(
-                        algorithm=algorithm_value,
-                    ),
-                )
-            assert (
-                str(exc_err.value)
-                == f"Dambreak field dambreakLevelsAndWidths can only be set when algorithm = 3, current value: {algorithm_value}."
-            )
-
-        def test_given_algorithm_value_3_returns_field_value(self):
-            field_value = "justAValue"
-            return_value = Dambreak.validate_dambreak_levels_and_widths(
-                field_value, dict(algorithm=3)
-            )
-            assert return_value == field_value
-
-        @pytest.mark.parametrize("algorithm_value", algorithm_values)
-        def test_given_none_field_value_and_algorithm_value_not_3_returns_field_value(
-            self, algorithm_value: int
-        ):
-            return_value = Dambreak.validate_dambreak_levels_and_widths(
-                None, dict(algorithm=algorithm_value)
-            )
-            assert return_value is None
-
-    class TestCheckLocation:
-        """
-        Wrapper to validate all paradigms of check_location
-        """
-
-        @pytest.mark.parametrize(
-            "dict_values",
-            [
-                pytest.param(
-                    dict(
-                        numcoordinates=2,
-                        xcoordinates=[0, 1],
-                        ycoordinates=[2, 3],
-                        waterlevelupstreamnodeid="anUpstreamNodeId",
-                        waterleveldownstreamnodeid="aDownstreamNodeId",
-                    ),
-                    id="With 2 coordinates and waterlevelupstreamnodeid and waterleveldownstreamnodeid",
-                ),
-                pytest.param(
-                    dict(
-                        numcoordinates=3,
-                        xcoordinates=[0, 1, 2],
-                        ycoordinates=[2, 3, 4],
-                        waterlevelupstreamnodeid="anUpstreamNodeId",
-                        waterleveldownstreamlocationx=3.4,
-                        waterleveldownstreamlocationy=4.5,
-                    ),
-                    id="With 3 coordinates and waterlevelupstreamnodeid and waterleveldownstreamlocationx and waterleveldownstreamlocationy",
-                ),
-                pytest.param(
-                    dict(
-                        polylinefile=Path(),
-                        waterlevelupstreamlocationx=1.2,
-                        waterlevelupstreamlocationy=2.3,
-                        waterleveldownstreamnodeid="aDownstreamNodeId",
-                    ),
-                    id="Empty path and waterlevelupstreamlocationx and waterlevelupstreamlocationy and waterleveldownstreamnodeid",
-                ),
-                pytest.param(
-                    dict(
-                        polylinefile=Path("aFilePath"),
-                        waterlevelupstreamlocationx=1.2,
-                        waterlevelupstreamlocationy=2.3,
-                        waterleveldownstreamlocationx=3.4,
-                        waterleveldownstreamlocationy=4.5,
-                    ),
-                    id="Path with file name and waterlevelupstreamlocationx and waterlevelupstreamlocationy and waterleveldownstreamlocationx and waterleveldownstreamlocationy",
-                ),
-            ],
-        )
-        def test_given_valid_values_returns_values(self, dict_values: dict):
-            # with patch("hydrolib.core.dflowfm.structure.models.Dambreak.model_dump") as mock_model_dump:
-            mock_dambreak = MagicMock(spec=Dambreak)
-            mock_dambreak.model_dump.return_value = dict_values
-            mock_dambreak.check_location_dambreak = Dambreak.check_location_dambreak.__get__(mock_dambreak, Dambreak)
-            mock_dambreak._validate_waterlevel_location = Dambreak._validate_waterlevel_location
-            assert mock_dambreak.check_location_dambreak()
-
-        @pytest.mark.parametrize(
-            "invalid_values, expected_err",
-            [
-                pytest.param(
-                    dict(
-                        numcoordinates=None,
-                        xcoordinates=None,
-                        ycoordinates=None,
-                        polylinefile=None,
-                        waterlevelupstreamnodeid="anUpstreamNodeId",
-                        waterleveldownstreamnodeid="aDownstreamNodeId",
-                    ),
-                    DambreakTestCases.check_location_err,
-                    id="Specified locations None",
-                ),
-                pytest.param(
-                    dict(
-                        waterleveldownstreamnodeid="aDownstreamNodeId",
-                        waterlevelupstreamnodeid=None,
-                    ),
-                    DambreakTestCases.check_upstream_waterlevel_location_err,
-                    id="No upstream water level locations specified",
-                ),
-                pytest.param(
-                    dict(
-                        waterleveldownstreamnodeid="aDownstreamNodeId",
-                        waterlevelupstreamlocationx=1.2,
-                    ),
-                    DambreakTestCases.check_upstream_waterlevel_location_err,
-                    id="Only upstream water level location x specified",
-                ),
-                pytest.param(
-                    dict(
-                        waterleveldownstreamnodeid="aDownstreamNodeId",
-                        waterlevelupstreamlocationy=2.3,
-                    ),
-                    DambreakTestCases.check_upstream_waterlevel_location_err,
-                    id="Only upstream water level location y specified",
-                ),
-                pytest.param(
-                    dict(
-                        waterleveldownstreamnodeid="aDownstreamNodeId",
-                        waterlevelupstreamnodeid="anUpstreamNodeId",
-                        waterlevelupstreamlocationx=1.2,
-                    ),
-                    DambreakTestCases.check_upstream_waterlevel_location_err,
-                    id="Upstream water level location node id and x specified",
-                ),
-                pytest.param(
-                    dict(
-                        waterleveldownstreamnodeid="aDownstreamNodeId",
-                        waterlevelupstreamnodeid="anUpstreamNodeId",
-                        waterlevelupstreamlocationy=2.3,
-                    ),
-                    DambreakTestCases.check_upstream_waterlevel_location_err,
-                    id="Upstream water level location node id and y specified",
-                ),
-                pytest.param(
-                    dict(
-                        waterleveldownstreamnodeid="aDownstreamNodeId",
-                        waterlevelupstreamnodeid="anUpstreamNodeId",
-                        waterlevelupstreamlocationx=1.2,
-                        waterlevelupstreamlocationy=2.3,
-                    ),
-                    DambreakTestCases.check_upstream_waterlevel_location_err,
-                    id="Upstream water level location node id, x and y specified",
-                ),
-                pytest.param(
-                    dict(
-                        waterlevelupstreamnodeid="anUpstreamNodeId",
-                        waterleveldownstreamnodeid=None,
-                    ),
-                    DambreakTestCases.check_downstream_waterlevel_location_err,
-                    id="No downstream water level locations specified",
-                ),
-                pytest.param(
-                    dict(
-                        waterlevelupstreamnodeid="anUpstreamNodeId",
-                        waterleveldownstreamlocationx=3.4,
-                    ),
-                    DambreakTestCases.check_downstream_waterlevel_location_err,
-                    id="Only downstream water level location x specified",
-                ),
-                pytest.param(
-                    dict(
-                        waterlevelupstreamnodeid="anUpstreamNodeId",
-                        waterleveldownstreamlocationy=4.5,
-                    ),
-                    DambreakTestCases.check_downstream_waterlevel_location_err,
-                    id="Only downstream water level location y specified",
-                ),
-                pytest.param(
-                    dict(
-                        waterlevelupstreamnodeid="anUpstreamNodeId",
-                        waterleveldownstreamnodeid="aDownstreamNodeId",
-                        waterleveldownstreamlocationx=3.4,
-                    ),
-                    DambreakTestCases.check_downstream_waterlevel_location_err,
-                    id="Downstream water level location node id and x specified",
-                ),
-                pytest.param(
-                    dict(
-                        waterlevelupstreamnodeid="anUpstreamNodeId",
-                        waterleveldownstreamnodeid="aDownstreamNodeId",
-                        waterleveldownstreamlocationy=4.5,
-                    ),
-                    DambreakTestCases.check_downstream_waterlevel_location_err,
-                    id="Downstream water level location node id and y specified",
-                ),
-                pytest.param(
-                    dict(
-                        waterlevelupstreamnodeid="anUpstreamNodeId",
-                        waterleveldownstreamnodeid="aDownstreamNodeId",
-                        waterleveldownstreamlocationx=3.4,
-                        waterleveldownstreamlocationy=4.5,
-                    ),
-                    DambreakTestCases.check_downstream_waterlevel_location_err,
-                    id="Downstream water level location node id, x and y specified",
-                ),
-            ],
-        )
-        def test_given_invalid_values_raises_expectation(
-            self, invalid_values: dict, expected_err: str, valid_dambreak_values: dict
-        ):
-            init_values = valid_dambreak_values
-            init_values.update(invalid_values)
-            with pytest.raises(ValueError) as exc_err:
-                _ = Dambreak(**init_values)
-            assert expected_err in str(exc_err.value)
-
-
-class TestOrifice:
-    """
-    Wrapper class to test all the methods and sublcasses in:
-    hydrolib.core.dflowfm.structure.models.py Orifice class.
-    """
-
-    def test_create_orifice(self):
-        structure = Orifice(**self._create_orifice_values())
-
-        assert structure.id == "structure_id"
-        assert structure.name == "structure_name"
-        assert structure.type == "orifice"
-        assert structure.branchid == "branch_id"
-        assert structure.chainage == pytest.approx(1.23)
-        assert structure.allowedflowdir == FlowDirection.positive
-        assert structure.crestlevel == pytest.approx(2.34)
-        assert structure.crestwidth == pytest.approx(3.45)
-        assert structure.gateloweredgelevel == pytest.approx(4.56)
-        assert structure.corrcoeff == pytest.approx(5.67)
-        assert structure.usevelocityheight == True
-        assert structure.uselimitflowpos == True
-        assert structure.limitflowpos == pytest.approx(6.78)
-        assert structure.uselimitflowneg == True
-        assert structure.limitflowneg == pytest.approx(7.89)
-
-    @pytest.mark.parametrize(
-        "limitflow, uselimitflow",
-        [
-            pytest.param("limitFlowPos", "useLimitFlowPos"),
-            pytest.param("limitFlowNeg", "useLimitFlowNeg"),
-        ],
-    )
-    def test_validate_limitflow(self, limitflow: str, uselimitflow: str):
-        values = self._create_orifice_values()
-        del values[limitflow.lower()]
-
-        with pytest.raises(ValueError):
-            Orifice(**values)
-
-
-    @pytest.mark.parametrize(
-        "input,expected",
-        _get_allowedflowdir_cases(),
-    )
-    def test_orifice_parses_flowdirection_case_insensitive(self, input, expected):
-        structure = Orifice(
-            allowedflowdir=input,
-            id="strucid",
-            branchid="branchid",
-            chainage="1",
-            crestlevel="1",
-            corrcoeff="1",
-            gateloweredgelevel="1",
-            usevelocityheight="0",
-            uselimitflowpos="0",
-            uselimitflowneg="0",
-        )
-
-        assert structure.allowedflowdir == expected
-
-    def test_optional_fields_have_correct_defaults(self):
-        structure = Orifice(**self._create_required_orifice_values())
-
-        assert structure.allowedflowdir == FlowDirection.both
-        assert structure.crestwidth == None
-        assert structure.uselimitflowpos == False
-        assert structure.limitflowpos == None
-        assert structure.uselimitflowneg == False
-        assert structure.limitflowneg == None
-
-    def _create_required_orifice_values(self) -> dict:
-        orifice_values = dict(
-            crestlevel="2.34",
-            gateloweredgelevel="4.56",
-            corrcoeff="5.67",
-            usevelocityheight="true",
-        )
-
-        orifice_values.update(create_structure_values("orifice"))
-
-        return orifice_values
-
-    def _create_orifice_values(self) -> dict:
-        orifice_values = dict(
-            allowedflowdir="positive",
-            crestwidth="3.45",
-            uselimitflowpos="true",
-            limitflowpos="6.78",
-            uselimitflowneg="true",
-            limitflowneg="7.89",
-        )
-
-        orifice_values.update(self._create_required_orifice_values())
-
-        return orifice_values
-
-
-class TestWeir:
-    def test_create_a_weir_from_scratch(self):
-        weir = Weir(
-            **self._create_weir_values(),
-            comments=Weir.Comments(
-                name="W stands for weir, 003 because we expect to have at most 999 weirs"
-            ),
-        )
-
-        assert weir.id == "structure_id"
-        assert weir.name == "structure_name"
-        assert weir.branchid == "branch_id"
-        assert weir.chainage == pytest.approx(1.23)
-        assert weir.allowedflowdir == FlowDirection.positive
-        assert weir.crestlevel == pytest.approx(2.34)
-        assert weir.crestwidth == pytest.approx(3.45)
-        assert weir.usevelocityheight == True
-        assert (
-            weir.comments.name
-            == "W stands for weir, 003 because we expect to have at most 999 weirs"
-        )
-
-        assert weir.comments.id == uniqueid_str
-
-    def test_id_comment_has_correct_default(self):
-        weir = Weir(**self._create_weir_values())
-
-        assert weir.comments.id == uniqueid_str
-
-    def test_add_comment_to_weir(self):
-        weir = Weir(**self._create_weir_values())
-
-        weir.comments.usevelocityheight = "a different value"
-        assert weir.comments.usevelocityheight == "a different value"
-
-    def test_weir_construction_with_parser(self):
-        parser = Parser(ParserConfig())
-
-        input_str = inspect.cleandoc(
-            """
-            [Structure]
-            id                = weir_id     # Unique structure id (max. 256 characters).
-            name              = weir        # Given name in the user interface.
-            branchId          = branch      # (optional) Branch on which the structure is located.
-            chainage          = 3.0         # (optional) Chainage on the branch (m)
-            type              = weir        # Structure type
-            allowedFlowDir    = positive    # Possible values: both, positive, negative, none.
-            crestLevel        = 10.5        # Crest level of weir (m AD)
-            crestWidth        =             # Width of weir (m)
-            useVelocityHeight = false       # Flag indicates whether the velocity height is to be calculated or not.
-            """
-        )
-
-        for line in input_str.splitlines():
-            parser.feed_line(line)
-
-        document = parser.finalize()
-
-        wrapper = WrapperTest[Weir].parse_obj({"val": document.sections[0]})
-        weir = wrapper.val
-
-        assert weir.id == "weir_id"
-        assert weir.name == "weir"
-        assert weir.branchid == "branch"
-        assert weir.chainage == pytest.approx(3.0)
-        assert weir.type == "weir"
-        assert weir.allowedflowdir == FlowDirection.positive
-        assert weir.crestlevel == pytest.approx(10.5)
-        assert weir.crestwidth is None
-        assert weir.usevelocityheight == False
-
-    def test_weir_comments_construction_with_parser(self):
-        parser = Parser(ParserConfig())
-
-        input_str = inspect.cleandoc(
-            """
-            [Structure]
-            id                = weir_id     
-            name              = weir        
-            branchId          = branch      
-            chainage          = 3.0         # My own special comment 1
-            type              = weir        
-            allowedFlowDir    = positive    
-            crestLevel        = 10.5        
-            crestWidth        =             
-            useVelocityHeight = false       # My own special comment 2
-            """
-        )
-
-        for line in input_str.splitlines():
-            parser.feed_line(line)
-
-        document = parser.finalize()
-
-        wrapper = WrapperTest[Weir].parse_obj({"val": document.sections[0]})
-        weir = wrapper.val
-
-        assert weir.comments.id is None
-        assert weir.comments.name is None
-        assert weir.comments.branchid is None
-        assert weir.comments.chainage == "My own special comment 1"
-        assert weir.comments.type is None
-        assert weir.comments.allowedflowdir is None
-        assert weir.comments.crestlevel is None
-        assert weir.comments.crestwidth is None
-        assert weir.comments.usevelocityheight == "My own special comment 2"
-
-    def test_weir_with_unknown_parameter_is_ignored(self):
-        parser = Parser(ParserConfig())
-
-        input_str = inspect.cleandoc(
-            """
-            [Structure]
-            id                = weir_id     # Unique structure id (max. 256 characters).
-            name              = weir        # Given name in the user interface.
-            branchId          = branch      # (optional) Branch on which the structure is located.
-            chainage          = 3.0         # (optional) Chainage on the branch (m)
-
-            # ----------------------------------------------------------------------
-            unknown           = 10.0        # A deliberately added unknown property
-            # ----------------------------------------------------------------------
-
-            type              = weir        # Structure type
-            allowedFlowDir    = positive    # Possible values: both, positive, negative, none.
-            crestLevel        = 10.5        # Crest level of weir (m AD)
-            crestWidth        =             # Width of weir (m)
-            useVelocityHeight = false       # Flag indicates whether the velocity height is to be calculated or not.
-            """
-        )
-
-        for line in input_str.splitlines():
-            parser.feed_line(line)
-
-        document = parser.finalize()
-
-        wrapper = WrapperTest[Weir].parse_obj({"val": document.sections[0]})
-        weir = wrapper.val
-
-        assert weir.dict().get("unknown") is None  # type: ignore
-
-        assert weir.id == "weir_id"
-        assert weir.name == "weir"
-        assert weir.branchid == "branch"
-        assert weir.chainage == pytest.approx(3.0)
-        assert weir.type == "weir"
-        assert weir.allowedflowdir == FlowDirection.positive
-        assert weir.crestlevel == pytest.approx(10.5)
-        assert weir.crestwidth is None
-        assert weir.usevelocityheight == False
-
-    @pytest.mark.parametrize(
-        "input,expected",
-        _get_allowedflowdir_cases(),
-    )
-    def test_weir_parses_flowdirection_case_insensitive(self, input, expected):
-        structure = Weir(
-            **self._create_required_weir_values(),
-            allowedflowdir=input,
-        )
-
-        assert structure.allowedflowdir == expected
-
-    def test_optional_fields_have_correct_defaults(self):
-        structure = Weir(**self._create_required_weir_values())
-
-        assert structure.allowedflowdir == FlowDirection.both
-        assert structure.crestwidth == None
-
-    def _create_required_weir_values(self) -> dict:
-        weir_values = dict(
-            crestlevel="2.34",
-            corrcoeff="5.67",
-            usevelocityheight="true",
-        )
-
-        weir_values.update(create_structure_values("weir"))
-        return weir_values
-
-    def _create_weir_values(self) -> dict:
-        weir_values = dict(
-            allowedflowdir="positive",
-            crestwidth="3.45",
-        )
-
-        weir_values.update(self._create_required_weir_values())
-        return weir_values
-
-
-class TestPump:
-    _pumplists = [
-        ("startlevelsuctionside", "numstages"),
-        ("stoplevelsuctionside", "numstages"),
-        ("startleveldeliveryside", "numstages"),
-        ("stopleveldeliveryside", "numstages"),
-        ("head", "numreductionlevels"),
-        ("reductionfactor", "numreductionlevels"),
-    ]
-
-    def test_create_a_pump_from_scratch(self):
-        pump = Pump(
-            **self._create_pump_values(),
-            comments=Pump.Comments(
-                name="P stands for pump, 003 because we expect to have at most 999 pumps"
-            ),
-        )
-
-        assert pump.id == "structure_id"
-        assert pump.name == "structure_name"
-        assert pump.branchid == "branch_id"
-        assert pump.capacity == pytest.approx(2.34)
-        assert pump.orientation == Orientation.negative
-        assert pump.numreductionlevels == 2
-        assert pump.head == [0, 2]
-        assert pump.reductionfactor == [0, 0.1]
-        assert (
-            pump.comments.name
-            == "P stands for pump, 003 because we expect to have at most 999 pumps"
-        )
-
-        assert pump.comments.id == uniqueid_str
-
-    def test_pump_construction_with_parser(self):
-        parser = Parser(ParserConfig())
-
-        input_str = inspect.cleandoc(
-            """
-            [Structure]
-            id                = pump_id     # Unique structure id (max. 256 characters).
-            name              = pump_nm     # Given name in the user interface.
-            branchId          = branch      # (optional) Branch on which the structure is located.
-            chainage          = 3.0         # (optional) Chainage on the branch (m)
-            type              = pump        # Structure type
-            orientation       = positive    # Possible values: positive, negative.
-            capacity          = 10.5        # Crest level of weir (m AD)
-            """
-        )
-
-        for line in input_str.splitlines():
-            parser.feed_line(line)
-
-        document = parser.finalize()
-
-        wrapper = WrapperTest[Pump].parse_obj({"val": document.sections[0]})
-        structure = wrapper.val
-
-        assert structure.id == "pump_id"
-        assert structure.name == "pump_nm"
-        assert structure.branchid == "branch"
-        assert structure.chainage == pytest.approx(3.0)
-        assert structure.type == "pump"
-        assert structure.orientation == Orientation.positive
-        assert structure.capacity == pytest.approx(10.5)
-
-    @pytest.mark.parametrize(
-        "input,expected",
-        _get_orientation_cases(),
-    )
-    def test_pump_parses_orientation_case_insensitive(self, input, expected):
-        structure = Pump(
-            **self._create_required_pump_values(),
-            orientation=input,
-        )
-
-        assert structure.orientation == expected
-
-    @pytest.mark.parametrize(
-        "listname,lengthname",
-        _pumplists,
-    )
-    def test_create_a_pump_with_wrong_list_lengths(
-        self, listname: str, lengthname: str
-    ):
-        """Creates a pump with one start/stop/reduction attribute with wrong list length
-        and checks for correct error detection."""
-        correctlength = 3  # just a test value
-        correctlist = list(range(correctlength))
-        listargs = {key: correctlist for (key, _) in self._pumplists}
-        listargs[listname] = list(
-            range(correctlength + 1)
-        )  # Intentional wrong list length
-        listargs[lengthname] = correctlength
-        listargs["controlside"] = "both"
-
-        with pytest.raises(ValueError) as error:
-
-            _ = Pump(
-                **self._create_required_pump_values(),
-                **listargs,
-            )
-
-        expected_message = f"Number of values for {listname} should be equal to the {lengthname} value."
-
-        assert expected_message in str(error.value)
-
-    @pytest.mark.parametrize(
-        "control_side,missing_list_name,present_list_names,should_raise_error",
-        [
-            ("suctionSide", "startlevelsuctionside", [], True),
-            (
-                "deliverySide",
-                "startlevelsuctionside",
-                ["startleveldeliveryside", "stopleveldeliveryside"],
-                False,
-            ),
-            ("both", "startlevelsuctionside", [], True),
-            ("suctionSide", "stoplevelsuctionside", [], True),
-            (
-                "deliverySide",
-                "stoplevelsuctionside",
-                ["startleveldeliveryside", "stopleveldeliveryside"],
-                False,
-            ),
-            ("both", "stoplevelsuctionside", [], True),
-            (
-                "suctionSide",
-                "startleveldeliveryside",
-                ["startlevelsuctionside", "stoplevelsuctionside"],
-                False,
-            ),
-            ("deliverySide", "startleveldeliveryside", [], True),
-            ("both", "startleveldeliveryside", [], True),
-            (
-                "suctionSide",
-                "stopleveldeliveryside",
-                ["startlevelsuctionside", "stoplevelsuctionside"],
-                False,
-            ),
-            ("deliverySide", "stopleveldeliveryside", [], True),
-            ("both", "stopleveldeliveryside", [], True),
-        ],
-    )
-    def test_dont_validate_unneeded_pump_lists(
-        self,
-        control_side: str,
-        missing_list_name: str,
-        present_list_names: List[str],
-        should_raise_error: bool,
-    ):
-        """Creates a pump with one particular list attribute missing
-        and checks for correct (i.e., no unneeded) error detection,
-        depending on the controlside attribute."""
-
-        values = self._create_required_pump_values()
-        values["id"] = "pump_controlside_" + (control_side or "none")
-
-        if control_side is not None:
-            values["controlside"] = control_side
-        else:
-            values.pop("controlside", None)
-
-        values["numstages"] = 1
-        values.pop(missing_list_name, None)
-        values.update({ln: [1.1] for ln in present_list_names})
-
-        if should_raise_error:
-            with pytest.raises(ValueError):
-                _ = Pump(**values)
-        else:
-            # Simply create the Pump and accept no Error raised.
-            _ = Pump(**values)
-
-    @pytest.mark.parametrize(
-        "listname,lengthname",
-        _pumplists,
-    )
-    def test_create_a_pump_with_missing_list(self, listname: str, lengthname: int):
-        """Creates a pump with one start/stop/reduction attribute list missing
-        and checks for correct error detection."""
-        correctlength = 3  # just a test value
-        correctlist = list(range(correctlength))
-        listargs = {key: correctlist for (key, _) in self._pumplists}
-        listargs.pop(listname)  # Remove one list argument
-        for _, _lengthname in self._pumplists:
-            listargs[_lengthname] = correctlength
-        listargs["controlside"] = "both"
-
-        with pytest.raises(ValueError) as error:
-
-            _ = Pump(
-                **self._create_required_pump_values(),
-                **listargs,
-            )
-
-        expected_message = (
-            f"List {listname} cannot be missing if {lengthname} is given."
-        )
-
-        assert expected_message in str(error.value)
-
-    def _create_required_pump_values(self) -> dict:
-        pump_values = dict(
-            capacity="2.34",
-        )
-
-        pump_values.update(create_structure_values("pump"))
-        return pump_values
-
-    def _create_pump_values(self) -> dict:
-        pump_values = dict(
-            orientation="negative",
-            numreductionlevels=2,
-            head=[0, 2],
-            reductionfactor=[0, 0.1],
-        )
-
-        pump_values.update(self._create_required_pump_values())
-        return pump_values
-
-
-class TestGeneralStructure:
-    def test_initialize_gateopeningwidth_with_timfile_initializes_timmodel(self):
-        gateloweredgelevel = test_input_dir / "tim" / "single_data_for_timeseries.tim"
-        values = self._create_general_structure_values()
-        values["gateloweredgelevel"] = gateloweredgelevel
-        structure = GeneralStructure(**values)
-
-        assert isinstance(structure.gateloweredgelevel, TimModel)
-
-    def test_initialize_gateopeningwidth_with_bcfile_initializes_forcingmodel(self):
-        gateloweredgelevel = (
-            test_input_dir
-            / "dflowfm_individual_files"
-            / "FlowFM_boundaryconditions2d_and_vectors.bc"
-        )
-        values = self._create_general_structure_values()
-        values["gateloweredgelevel"] = gateloweredgelevel
-        structure = GeneralStructure(**values)
-
-        assert isinstance(structure.gateloweredgelevel, ForcingModel)
-
-    def test_initialize_gateopeningwidth_with_float_value_initializes_float(self):
-        gateloweredgelevel = 123.456
-        values = self._create_general_structure_values()
-        values["gateloweredgelevel"] = gateloweredgelevel
-        structure = GeneralStructure(**values)
-
-        assert isinstance(structure.gateloweredgelevel, float)
-
-    def test_create_a_general_structure_from_scratch(self):
-        name_comment = "Generic name comment"
-        struct = GeneralStructure(
-            **self._create_general_structure_values(),
-            comments=GeneralStructure.Comments(name=name_comment),
-        )
-
-        assert struct.id == "structure_id"
-        assert struct.name == "structure_name"
-        assert struct.branchid == "branch_id"
-        assert struct.chainage == pytest.approx(1.23)
-
-        assert struct.allowedflowdir == FlowDirection.positive
-        assert struct.upstream1width == pytest.approx(1.0)
-        assert struct.upstream1level == pytest.approx(2.0)
-        assert struct.upstream2width == pytest.approx(3.0)
-        assert struct.upstream2level == pytest.approx(4.0)
-        assert struct.crestwidth == pytest.approx(5.0)
-        assert struct.crestlevel == pytest.approx(6.0)
-        assert struct.crestlength == pytest.approx(7.0)
-        assert struct.downstream1width == pytest.approx(8.0)
-        assert struct.downstream1level == pytest.approx(9.0)
-        assert struct.downstream2width == pytest.approx(9.1)
-        assert struct.downstream2level == pytest.approx(9.2)
-        assert struct.gateloweredgelevel == pytest.approx(9.3)
-        assert struct.posfreegateflowcoeff == pytest.approx(9.4)
-        assert struct.posdrowngateflowcoeff == pytest.approx(9.5)
-        assert struct.posfreeweirflowcoeff == pytest.approx(9.6)
-        assert struct.posdrownweirflowcoeff == pytest.approx(9.7)
-        assert struct.poscontrcoeffreegate == pytest.approx(9.8)
-        assert struct.negfreegateflowcoeff == pytest.approx(9.9)
-        assert struct.negdrowngateflowcoeff == pytest.approx(8.1)
-        assert struct.negfreeweirflowcoeff == pytest.approx(8.2)
-        assert struct.negdrownweirflowcoeff == pytest.approx(8.3)
-        assert struct.negcontrcoeffreegate == pytest.approx(8.4)
-        assert struct.extraresistance == pytest.approx(8.5)
-        assert struct.gateheight == pytest.approx(8.6)
-        assert struct.gateopeningwidth == pytest.approx(8.7)
-        assert (
-            struct.gateopeninghorizontaldirection
-            == GateOpeningHorizontalDirection.from_left
-        )
-        assert struct.usevelocityheight == False
-
-        assert struct.comments is not None
-        assert struct.comments.name == name_comment
-
-        assert struct.comments.id == uniqueid_str
-
-    def test_id_comment_has_correct_default(self):
-        struct = GeneralStructure(**self._create_general_structure_values())
-        assert struct.comments is not None
-        assert struct.comments.id == uniqueid_str
-
-    def test_add_comment_to_general_structure(self):
-        struct = GeneralStructure(**self._create_general_structure_values())
-
-        assert struct.comments is not None
-
-        new_comment = "a very different value"
-
-        struct.comments.usevelocityheight = new_comment
-        assert struct.comments.usevelocityheight == new_comment
-
-    def test_general_structure_construction_with_parser(self):
-        parser = Parser(ParserConfig())
-
-        input_str = inspect.cleandoc(
-            """
-            [Structure]
-            id                             = potato_id        # Unique structure id (max. 256 characters).
-            name                           = structure_potato # Given name in the user interface.
-            type                           = generalStructure # Structure type; must read generalStructure
-            branchId                       = branch           # Branch on which the structure is located.
-            chainage                       = 3.53             # Chainage on the branch (m).
-            allowedFlowDir                 = positive         # Possible values: both, positive, negative, none.
-            upstream1Width                 = 11.0             # w_u1 [m]
-            upstream1Level                 = 12.0             # z_u1 [m AD]
-            upstream2Width                 = 13.0             # w_u2 [m]
-            upstream2Level                 = 14.0             # z_u2 [m D]
-            crestWidth                     = 15.0             # w_s [m]
-            crestLevel                     = 16.0             # z_s [m AD]
-            crestLength                    = 17.0             # The crest length across the general structure [m]. When the crest length > 0, the extra resistance for this structure will be ls * g/(C2 * waterdepth)
-            downstream1Width               = 18.0             # w_d1 [m]
-            downstream1Level               = 19.0             # z_d1 [m AD]
-            downstream2Width               = 19.1             # w_d1 [m]
-            downstream2Level               = 19.2             # z_d1 [m AD]
-            gateLowerEdgeLevel             = 19.3             # Position of gate door’s lower edge [m AD]
-            posFreeGateFlowCoeff           = 19.4             # Positive free gate flow corr.coeff. cgf [-]
-            posDrownGateFlowCoeff          = 19.5             # Positive drowned gate flow corr.coeff. cgd [-]
-            posFreeWeirFlowCoeff           = 19.6             # Positive free weir flow corr.coeff. cwf [-]
-            posDrownWeirFlowCoeff          = 19.7             # Positive drowned weir flow corr.coeff. cwd [-]
-            posContrCoefFreeGate           = 19.8             # Positive gate flow contraction coefficient µgf [-]
-            negFreeGateFlowCoeff           = 19.9             # Negative free gate flow corr.coeff. cgf [-]
-            negDrownGateFlowCoeff          = 18.1             # Negative drowned gate flow corr.coeff. cgd [-]
-            negFreeWeirFlowCoeff           = 18.2             # Negative free weir flow corr.coeff. cwf [-]
-            negDrownWeirFlowCoeff          = 18.3             # Negative drowned weir flow corr.coeff. cwd [-]
-            negContrCoefFreeGate           = 18.4             # Negative gate flow contraction coefficient mu gf [-]
-            extraResistance                = 18.5             # Extra resistance [-]
-            gateHeight                     = 18.6
-            gateOpeningWidth               = 18.7             # Opening width between gate doors [m], should be smaller than (or equal to) crestWidth
-            gateOpeningHorizontalDirection = fromLeft         # Horizontal opening direction of gate door[s]. Possible values are: symmetric, fromLeft, fromRight
-            useVelocityHeight              = 0                # Flag indicates whether the velocity height is to be calculated or not
-            """
-        )
-
-        for line in input_str.splitlines():
-            parser.feed_line(line)
-
-        document = parser.finalize()
-
-        wrapper = WrapperTest[GeneralStructure].parse_obj({"val": document.sections[0]})
-        struct = wrapper.val
-
-        assert struct.id == "potato_id"
-        assert struct.name == "structure_potato"
-        assert struct.branchid == "branch"
-        assert struct.chainage == pytest.approx(3.53)
-        assert struct.type == "generalStructure"
-        assert struct.allowedflowdir == FlowDirection.positive
-        assert struct.upstream1width == pytest.approx(11.0)
-        assert struct.upstream1level == pytest.approx(12.0)
-        assert struct.upstream2width == pytest.approx(13.0)
-        assert struct.upstream2level == pytest.approx(14.0)
-        assert struct.crestwidth == pytest.approx(15.0)
-        assert struct.crestlevel == pytest.approx(16.0)
-        assert struct.crestlength == pytest.approx(17.0)
-        assert struct.downstream1width == pytest.approx(18.0)
-        assert struct.downstream1level == pytest.approx(19.0)
-        assert struct.downstream2width == pytest.approx(19.1)
-        assert struct.downstream2level == pytest.approx(19.2)
-        assert struct.gateloweredgelevel == pytest.approx(19.3)
-        assert struct.posfreegateflowcoeff == pytest.approx(19.4)
-        assert struct.posdrowngateflowcoeff == pytest.approx(19.5)
-        assert struct.posfreeweirflowcoeff == pytest.approx(19.6)
-        assert struct.posdrownweirflowcoeff == pytest.approx(19.7)
-        assert struct.poscontrcoeffreegate == pytest.approx(19.8)
-        assert struct.negfreegateflowcoeff == pytest.approx(19.9)
-        assert struct.negdrowngateflowcoeff == pytest.approx(18.1)
-        assert struct.negfreeweirflowcoeff == pytest.approx(18.2)
-        assert struct.negdrownweirflowcoeff == pytest.approx(18.3)
-        assert struct.negcontrcoeffreegate == pytest.approx(18.4)
-        assert struct.extraresistance == pytest.approx(18.5)
-        assert struct.gateheight == pytest.approx(18.6)
-        assert struct.gateopeningwidth == pytest.approx(18.7)
-        assert (
-            struct.gateopeninghorizontaldirection
-            == GateOpeningHorizontalDirection.from_left
-        )
-        assert struct.usevelocityheight == False
-
-    def test_weir_comments_construction_with_parser(self):
-        parser = Parser(ParserConfig())
-
-        input_str = inspect.cleandoc(
-            """
-            [Structure]
-            id                             = potato_id        
-            name                           = structure_potato 
-            type                           = generalStructure 
-            branchId                       = branch           
-            chainage                       = 3.53   # My own special comment 1
-            allowedFlowDir                 = positive         
-            upstream1Width                 = 11.0             
-            upstream1Level                 = 12.0             
-            upstream2Width                 = 13.0             
-            upstream2Level                 = 14.0             
-            crestWidth                     = 15.0             
-            crestLevel                     = 16.0             
-            crestLength                    = 17.0             
-            downstream1Width               = 18.0             
-            downstream1Level               = 19.0             
-            downstream2Width               = 19.1             
-            downstream2Level               = 19.2             
-            gateLowerEdgeLevel             = 19.3             
-            posFreeGateFlowCoeff           = 19.4             
-            posDrownGateFlowCoeff          = 19.5             
-            posFreeWeirFlowCoeff           = 19.6             
-            posDrownWeirFlowCoeff          = 19.7             
-            posContrCoefFreeGate           = 19.8             
-            negFreeGateFlowCoeff           = 19.9             
-            negDrownGateFlowCoeff          = 18.1             
-            negFreeWeirFlowCoeff           = 18.2             
-            negDrownWeirFlowCoeff          = 18.3             
-            negContrCoefFreeGate           = 18.4             
-            extraResistance                = 18.5             
-            gateHeight                     = 18.6
-            gateOpeningWidth               = 18.7              
-            gateOpeningHorizontalDirection = fromLeft         
-            useVelocityHeight              = 0    # My own special comment 2            
-            """
-        )
-
-        for line in input_str.splitlines():
-            parser.feed_line(line)
-
-        document = parser.finalize()
-
-        wrapper = WrapperTest[GeneralStructure].parse_obj({"val": document.sections[0]})
-        struct = wrapper.val
-
-        assert struct.comments is not None
-        assert struct.comments.id is None
-        assert struct.comments.name is None
-        assert struct.comments.branchid is None
-        assert struct.comments.chainage == "My own special comment 1"
-        assert struct.comments.type is None
-
-        assert struct.comments.allowedflowdir is None
-        assert struct.comments.upstream1width is None
-        assert struct.comments.upstream1level is None
-        assert struct.comments.upstream2width is None
-        assert struct.comments.upstream2level is None
-        assert struct.comments.crestwidth is None
-        assert struct.comments.crestlevel is None
-        assert struct.comments.crestlength is None
-        assert struct.comments.downstream1width is None
-        assert struct.comments.downstream1level is None
-        assert struct.comments.downstream2width is None
-        assert struct.comments.downstream2level is None
-        assert struct.comments.gateloweredgelevel is None
-        assert struct.comments.posfreegateflowcoeff is None
-        assert struct.comments.posdrowngateflowcoeff is None
-        assert struct.comments.posfreeweirflowcoeff is None
-        assert struct.comments.posdrownweirflowcoeff is None
-        assert struct.comments.poscontrcoeffreegate is None
-        assert struct.comments.negfreegateflowcoeff is None
-        assert struct.comments.negdrowngateflowcoeff is None
-        assert struct.comments.negfreeweirflowcoeff is None
-        assert struct.comments.negdrownweirflowcoeff is None
-        assert struct.comments.negcontrcoeffreegate is None
-        assert struct.comments.extraresistance is None
-        assert struct.comments.gateheight is None
-        assert struct.comments.gateopeningwidth is None
-        assert struct.comments.usevelocityheight == "My own special comment 2"
-
-    def test_general_structure_with_unknown_parameter_is_ignored(self):
-        parser = Parser(ParserConfig())
-
-        input_str = inspect.cleandoc(
-            """
-            [Structure]
-            id                             = id        
-            name                           = extravagante_waarde
-
-            # ----------------------------------------------------------------------
-            unknown           = 10.0        # A deliberately added unknown property
-            # ----------------------------------------------------------------------
-
-            type                           = generalStructure 
-            branchId                       = stump
-            chainage                       = 13.53  
-            allowedFlowDir                 = positive         
-            upstream1Width                 = 111.0             
-            upstream1Level                 = 112.0             
-            upstream2Width                 = 113.0             
-            upstream2Level                 = 114.0             
-            crestWidth                     = 115.0             
-            crestLevel                     = 116.0             
-            crestLength                    = 117.0             
-            downstream1Width               = 118.0             
-            downstream1Level               = 119.0             
-            downstream2Width               = 119.1             
-            downstream2Level               = 119.2             
-            gateLowerEdgeLevel             = 119.3             
-            posFreeGateFlowCoeff           = 119.4             
-            posDrownGateFlowCoeff          = 119.5             
-            posFreeWeirFlowCoeff           = 119.6             
-            posDrownWeirFlowCoeff          = 119.7             
-            posContrCoefFreeGate           = 119.8             
-            negFreeGateFlowCoeff           = 119.9             
-            negDrownGateFlowCoeff          = 118.1             
-            negFreeWeirFlowCoeff           = 118.2             
-            negDrownWeirFlowCoeff          = 118.3             
-            negContrCoefFreeGate           = 118.4             
-            extraResistance                = 118.5             
-            gateHeight                     = 118.6
-            gateOpeningWidth               = 118.7              
-            gateOpeningHorizontalDirection = fromRight
-            useVelocityHeight              = 0
-            """
-        )
-
-        for line in input_str.splitlines():
-            parser.feed_line(line)
-
-        document = parser.finalize()
-
-        wrapper = WrapperTest[GeneralStructure].parse_obj({"val": document.sections[0]})
-        struct = wrapper.val
-
-        assert struct.dict().get("unknown") is None  # type: ignore
-
-        assert struct.id == "id"
-        assert struct.name == "extravagante_waarde"
-        assert struct.branchid == "stump"
-        assert struct.chainage == pytest.approx(13.53)
-        assert struct.type == "generalStructure"
-        assert struct.allowedflowdir == FlowDirection.positive
-        assert struct.upstream1width == pytest.approx(111.0)
-        assert struct.upstream1level == pytest.approx(112.0)
-        assert struct.upstream2width == pytest.approx(113.0)
-        assert struct.upstream2level == pytest.approx(114.0)
-        assert struct.crestwidth == pytest.approx(115.0)
-        assert struct.crestlevel == pytest.approx(116.0)
-        assert struct.crestlength == pytest.approx(117.0)
-        assert struct.downstream1width == pytest.approx(118.0)
-        assert struct.downstream1level == pytest.approx(119.0)
-        assert struct.downstream2width == pytest.approx(119.1)
-        assert struct.downstream2level == pytest.approx(119.2)
-        assert struct.gateloweredgelevel == pytest.approx(119.3)
-        assert struct.posfreegateflowcoeff == pytest.approx(119.4)
-        assert struct.posdrowngateflowcoeff == pytest.approx(119.5)
-        assert struct.posfreeweirflowcoeff == pytest.approx(119.6)
-        assert struct.posdrownweirflowcoeff == pytest.approx(119.7)
-        assert struct.poscontrcoeffreegate == pytest.approx(119.8)
-        assert struct.negfreegateflowcoeff == pytest.approx(119.9)
-        assert struct.negdrowngateflowcoeff == pytest.approx(118.1)
-        assert struct.negfreeweirflowcoeff == pytest.approx(118.2)
-        assert struct.negdrownweirflowcoeff == pytest.approx(118.3)
-        assert struct.negcontrcoeffreegate == pytest.approx(118.4)
-        assert struct.extraresistance == pytest.approx(118.5)
-        assert struct.gateheight == pytest.approx(118.6)
-        assert struct.gateopeningwidth == pytest.approx(118.7)
-        assert (
-            struct.gateopeninghorizontaldirection
-            == GateOpeningHorizontalDirection.from_right
-        )
-        assert struct.usevelocityheight == False
-
-    def _create_required_general_structure_values(self) -> dict:
-        general_structure_values = dict()
-        general_structure_values.update(create_structure_values("generalStructure"))
-        return general_structure_values
-
-    def _create_general_structure_values(self) -> dict:
-        general_structure_values = dict(
-            allowedflowdir=FlowDirection.positive,
-            upstream1width=1.0,
-            upstream1level=2.0,
-            upstream2width=3.0,
-            upstream2level=4.0,
-            crestwidth=5.0,
-            crestlevel=6.0,
-            crestlength=7.0,
-            downstream1width=8.0,
-            downstream1level=9.0,
-            downstream2width=9.1,
-            downstream2level=9.2,
-            gateloweredgelevel=9.3,
-            posfreegateflowcoeff=9.4,
-            posdrowngateflowcoeff=9.5,
-            posfreeweirflowcoeff=9.6,
-            posdrownweirflowcoeff=9.7,
-            poscontrcoeffreegate=9.8,
-            negfreegateflowcoeff=9.9,
-            negdrowngateflowcoeff=8.1,
-            negfreeweirflowcoeff=8.2,
-            negdrownweirflowcoeff=8.3,
-            negcontrcoeffreegate=8.4,
-            extraresistance=8.5,
-            gateheight=8.6,
-            gateopeningwidth=8.7,
-            gateopeninghorizontaldirection=GateOpeningHorizontalDirection.from_left,
-            usevelocityheight=False,
-        )
-
-        general_structure_values.update(
-            self._create_required_general_structure_values()
-        )
-        return general_structure_values
-
-
-class TestCulvert:
-    @pytest.mark.parametrize(
-        "missing_field",
-        ["valveopeningheight", "numlosscoeff", "relopening", "losscoeff"],
-    )
-    def test_validate_required_fields_based_on_valveonoff(self, missing_field: str):
-        values = self._create_culvert_values(valveonoff=True)
-        del values[missing_field]
-
-        with pytest.raises(ValueError) as error:
-            Culvert(**values)
-
-        expected_message = f"{missing_field} should be provided when valveonoff is True"
-        assert expected_message in str(error.value)
-
-    def test_validate_default_subtype(self):
-        culvert = Culvert(**self._create_culvert_values(valveonoff=False))
-
-        assert culvert.subtype == CulvertSubType.culvert
-
-    def test_validate_bendlosscoeff_required_when_invertedsiphon_subtype(self):
-        values = self._create_culvert_values(valveonoff=False)
-        values["subtype"] = CulvertSubType.invertedSiphon
-
-        with pytest.raises(ValueError) as error:
-            Culvert(**values)
-
-        expected_message = (
-            "bendlosscoeff should be provided when subtype is invertedSiphon"
-        )
-        assert expected_message in str(error.value)
-
-    def test_validate_bendlosscoeff_forbidden_when_culvert_subtype(self):
-        values = self._create_culvert_values(valveonoff=False)
-        values["subtype"] = CulvertSubType.culvert
-        values["bendlosscoeff"] = 0.1
-
-        with pytest.raises(ValueError) as error:
-            Culvert(**values)
-
-        expected_message = "bendlosscoeff is forbidden when subtype is culvert"
-        assert expected_message in str(error.value)
-
-    def _create_culvert_values(self, valveonoff: bool):
-        values = create_structure_values("culvert")
-        values.update(
-            dict(
-                allowedflowdir="both",
-                leftlevel="1.23",
-                rightlevel="2.34",
-                csdefid="cs_def_id",
-                length="3.45",
-                inletlosscoeff="4.56",
-                outletlosscoeff="5.67",
-                valveonoff=valveonoff,
-                bedfrictiontype=FrictionType.whitecolebrook,
-                bedfriction="4.32",
-            )
-        )
-
-        if valveonoff:
-            values.update(
-                dict(
-                    valveopeningheight="6.78",
-                    numlosscoeff="10",
-                    relopening="7.89 9.87 8.76",
-                    losscoeff="7.65 6.54 5.43",
-                )
-            )
-
-        return values
-
-
-class TestLongCulvert:
-    @pytest.fixture
-    def longculvert_values(self) -> Dict[str, Any]:
-        return {
-            "id": "lc1",
-            "name": "Long Culvert 1",
-            "type": "longCulvert",
-            "branchid": "branch_id",
-            "chainage": 3.53,
-            "numcoordinates": 2,
-            "xcoordinates": [6.515339, 44.636787],
-            "ycoordinates": [25.151608, 25.727361],
-            "zcoordinates": [-0.3, -0.3],
-            "width": 0.4,
-            "height": 0.2,
-            "frictiontype": FrictionType.manning,
-            "frictionvalue": 0.035,
-            "valverelativeopening": 1.0,
-        }
-
-    def test_create_longculvert_minimal(self, longculvert_values: Dict[str, Any]):
-        lc = LongCulvert(**longculvert_values)
-        assert lc.id == "lc1"
-        assert lc.name == "Long Culvert 1"
-        assert lc.type == "longCulvert"
-        assert lc.numcoordinates == 2
-        assert lc.xcoordinates == [6.515339, 44.636787]
-        assert lc.ycoordinates == [25.151608, 25.727361]
-        assert lc.zcoordinates == [-0.3, -0.3]
-        assert lc.width == pytest.approx(0.4)
-        assert lc.height == pytest.approx(0.2)
-        assert lc.frictiontype == FrictionType.manning
-        assert lc.frictionvalue == pytest.approx(0.035)
-        assert lc.valverelativeopening == pytest.approx(1.0)
-
-    @pytest.mark.parametrize(
-        "missing_field",
-        ["numcoordinates", "xcoordinates", "ycoordinates"],
-    )
-    def test_missing_coordinates_raises(
-        self, missing_field, longculvert_values: Dict[str, Any]
-    ):
-        del longculvert_values[missing_field]
-        with pytest.raises(ValueError):
-            LongCulvert(**longculvert_values)
-
-    def test_invalid_coordinates_length_raises(
-        self, longculvert_values: Dict[str, Any]
-    ):
-        longculvert_values["numcoordinates"] = 3
-        with pytest.raises(ValueError):
-            LongCulvert(**longculvert_values)
-
-    def test_invalid_zcoordinates_length_raises(
-        self, longculvert_values: Dict[str, Any]
-    ):
-        longculvert_values["zcoordinates"] = [-0.3, -0.3, -0.4]
-        with pytest.raises(ValueError):
-            LongCulvert(**longculvert_values)
-
-    def test_create_longculvert_without_zcoordinates(
-        self, longculvert_values: Dict[str, Any]
-    ):
-        del longculvert_values["zcoordinates"]
-
-        lc = LongCulvert(**longculvert_values)
-        assert lc.zcoordinates is None
